Support emterpreter sync, enabling things which failed before.
Restructuring DOSBox to fully work with the Emscripten main loop
would be difficult because of the many places which can lead to
nested inner loops. Emscripten emterpreter sync support allows
these things to work as-is, but emterpreter is slow. This only
uses the emterpreter for a few functions which may need to be
interrupted in common use. Sleep is not allowed in CPU exceptions
and interrupts, and those will abort emulation with a timeout
error if they don't finish promptly.

Give the --enable-sync option to ./configure to enable this. The
interactive DOS prompt, batch files, and running programs from
other programs should all work with this option enabled.