Partially fix splash screen, enabling SDL 2 WebGL texture output
The splash screen was running but not getting displayed. This
created a 2D canvas context which prevented subsequent creation
of a 3D context for texture output.

The splash screen code is now disabled when not using emterpreter
sync and skipped if using "texture" or "texturenb" output. It
works if using emterpreter sync and surface output.