Disable memory growth because of its performance penalty
DOSBox should not need memory growth because DOS programs do
not need much memory. Enabling it gives a noticeable performance
penalty in Firefox 35 and a severe performance penalty in the
Chrome 42 dev branch. Recently, Emscripten started enabling it
by default, so a flag needs to be added to disable it.